requirements.txt
tensorflow
python-dotenv==1.0.1